Rename utils and add test for `pxStringToNumber`
