+++ libs/@guardian/source-foundations/src/utils/convert-value.ts
@@ -1,4 +1,3 @@
 export const rootPixelFontSize = 16;
 export const pxToRem = (px: number): number => px / rootPixelFontSize;
-
 export const pxStringToNumber = (px: string): number => Number(px.slice(0, -2));
